% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_functions.R
\name{boxplot_mda}
\alias{boxplot_mda}
\title{The boxplot_mda() function combines scaled vectors of the relevant factor loadings and boxplots of diminesion scores.}
\usage{
boxplot_mda(mda_data, n_factor = 1)
}
\arguments{
\item{mda_data}{An mda data.frame produced by the mda_loadings() function.}

\item{n_factor}{The factor to be plotted.}
}
\value{
A combined plot of scaled vectors and boxplots.
}
\description{
The boxplot_mda() function combines scaled vectors of the relevant factor loadings and boxplots of diminesion scores.
}
