% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_functions.R
\name{heatmap_mda}
\alias{heatmap_mda}
\title{The heatmap_mda() function combines a stick plot with a heat map of the relevant factor loadings.}
\usage{
heatmap_mda(mda_data, n_factor = 1)
}
\arguments{
\item{mda_data}{An mda data.frame produced by the mda_loadings() function.}

\item{n_factor}{The factor to be plotted.}
}
\value{
A comined stick plot and heat map.
}
\description{
The heatmap_mda() function combines a stick plot with a heat map of the relevant factor loadings.
}
