% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{micusp_biber}
\alias{micusp_biber}
\title{MICUSP corpus tagged with pseudobibeR features}
\format{
A data frame containing:
\describe{
\item{doc_id}{Document ID (from MICUSP)}
\item{remaining columns}{Rate of feature use per 1,000 tokens. See \code{pseudobibeR::biber()} for description of each feature.}
}
}
\source{
Michigan Corpus of Upper-Level Student Papers,
\url{https://elicorpora.info/main}, tagged with the pseudobibeR package.
}
\usage{
micusp_biber
}
\description{
The Michigan Corpus of Upper-Level Student Papers (MICUSP) contains 828
student papers. Here each document is tagged with Biber features using the
pseudobibeR package. Type-to-token ratio is calculated using the moving
average type-to-token ratio (MATTR).
}
\keyword{datasets}
