% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_functions.R
\name{stickplot_mda}
\alias{stickplot_mda}
\alias{heatmap_mda}
\title{Plots of MDA factor group means and loadings}
\usage{
stickplot_mda(mda_data, n_factor = 1)

heatmap_mda(mda_data, n_factor = 1)
}
\arguments{
\item{mda_data}{An mda data frame produced by the \code{mda_loadings()} function.}

\item{n_factor}{Index of the factor to be plotted.}
}
\value{
ggplot object
}
\description{
Stick plots show each group's mean loading along a factor, plotted along a
positive/negative cline. Heatmaps show each variable's loading on a factor.
\code{stickplot_mda()} produces just a stick plot, while \code{heatmap_mda()} places a
heatmap alongside the stick plot.
}
\seealso{
\code{\link[=boxplot_mda]{boxplot_mda()}}
}
