% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_functions.R
\name{boxplot_mda}
\alias{boxplot_mda}
\title{Create boxplot for multi-dimensional analysis}
\usage{
boxplot_mda(mda_data, n_factor = 1)
}
\arguments{
\item{mda_data}{An mda data.frame produced by the \code{mda_loadings()} function.}

\item{n_factor}{The factor to be plotted.}
}
\value{
A combined plot of scaled vectors and boxplots.
}
\description{
Combine scaled vectors of the relevant factor loadings and boxplots of
dimension scores.
}
\seealso{
\code{\link[=stickplot_mda]{stickplot_mda()}}
}
