% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_functions.R
\name{heatmap_mda}
\alias{heatmap_mda}
\title{Create heatmap plots for multi-dimensional analysis}
\usage{
heatmap_mda(mda_data, n_factor = 1)
}
\arguments{
\item{mda_data}{An mda data.frame produced by the \code{mda_loadings()} function.}

\item{n_factor}{The factor to be plotted.}
}
\value{
A combined stick plot and heat map.
}
\description{
Combine a stick plot with a heat map of the relevant factor loadings.
}
