% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_functions.R
\name{mda_loadings}
\alias{mda_loadings}
\title{Multi-Dimensional Analysis is a statistical procedure developed Biber and is commonly used in descriptions
of language as it varies by genre, register, and task. The procedure is a specific application
factor analysis, which is used as the basis for calculating a 'dimension score' for each text.}
\usage{
mda_loadings(data_frame, n_factors, cor_min = 0.2, threshold = 0.35)
}
\arguments{
\item{data_frame}{A data.frame containing 1 categorical (factor) variable and continuous (numeric) variables.}

\item{n_factors}{The number of factors to be calculated in the factor analysis.}

\item{cor_min}{The correlation threshold for including variables in the factor analysis.}

\item{threshold}{A value indicating the threshold at which variables should be included in dimension score calculations (the default is .35).}
}
\value{
An mda data structure containing scores, means by group, and factor loadings
}
\description{
The function mda_loadings() returns a data.frame of dimension scores with the means for each category
and the factor loadings accessible as attributes. Calculating MDA requires a data.frame containing
a column with a categorical variable (formatted as a factor) and more than 2 continuous, numeric variables.
}
