% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_functions.R
\name{stickplot_mda}
\alias{stickplot_mda}
\title{Create stick plots for multi-dimensional analysis}
\usage{
stickplot_mda(mda_data, n_factor = 1)
}
\arguments{
\item{mda_data}{An mda data.frame produced by the mda_loadings() function.}

\item{n_factor}{The factor to be plotted.}
}
\value{
A stick plot showing category means along a positive/negative cline.
}
\description{
A simple function for producing the stick plots that are common in
visualizing the location of category means along a given dimension.
}
